requirements.txt
httpx==0.28.1
beautifulsoup4==4.14.2
supabase==2.10.0
lxml==5.3.0
psutil==6.1.0
python-dotenv==1.0.1